+++ src/nodejs/qm/qm_nodejs.h
@@ -8,6 +8,16 @@
 #ifndef QMINER_QM_NODEJS
 #define QMINER_QM_NODEJS
 
+#include <node.h>
+#include <node_object_wrap.h>
+#include <qminer.h>
+#include "../la/la_nodejs.h"
+#include "../fs/fs_nodejs.h"
+#include "../nodeutil.h"
+
+#include "qm_nodejs_streamaggr.h"
+#include "qm_nodejs_store.h"
+
 ///////////////////////////////
 // NodeJs QMiner.
 // A factory of base objects
@@ -66,6 +76,14 @@ private:
     */
     //# exports.verbosity = function (level) { }
     JsDeclareFunction(verbosity);
+
+    /**
+    * Returns an JSON with two properties: "byClass" and "total". The "byClass" value is a JSON where
+    * each key is a class ID and each value is of the form { newFromCpp: number, newFromJs: number, destructorCalls: number}
+    * and the value of "total" is of the same form (aggregated over "byClass")    
+    */
+    //# exports.stats = function () { }
+    JsDeclareFunction(stats);
     
     /**
     * @typedef {Object} QMinerFlags
@@ -164,7 +182,9 @@ private:
 * <br>8. `'float_pair'` - A pair of floats, useful for storing geo coordinates,
 * <br>9. `'float_v'` - Array of floats,
 * <br>10. `'datetime'` - Date and time format, stored in a form of miliseconds since 1600,
-* <br>11. `'num_sp_v'` - Array of [`int`, `float`] pairs. See constructor array for {@link module:la.SparseVector}.
+* <br>11. `'num_sp_v'` - Array of [`int`, `float`] pairs. See constructor array for {@link module:la.SparseVector},
+* <br>12. `'json'` - this field can be an arbitrary object and will be internally serialized into string using JSON notation,
+* <br>13. `'blob'` - Binary buffer, used for storing binary data,
 * @property {boolean} [primary=false] - Field which can be used to identify record. There can be only one primary field in a store. There can be at most one record for each value of the primary field. Currently following fields can be marked as primary: `int`, `uint64`, `string`, `float`, `datetime`. Primary fields of type `string` are also used for record querying using {@link module:qm.Store#recordByName}.
 * @property {boolean} [null=false] - When set to true, null is a possible value for a field (allow missing values).
 * @property {string} [store='memory'] - Defines where to store the field. Possible options 
@@ -411,6 +431,7 @@ class TNodeJsBase : public node::ObjectWrap {
     friend class TNodeJsUtil;
 private:
     static v8::Persistent<v8::Function> Constructor;
+    ~TNodeJsBase() { TNodeJsUtil::ObjNameH.GetDat(GetClassId()).Val3++; TNodeJsUtil::ObjCount.Val3++; }
 public:
     static const int MAX_BASES;
     static void Init(v8::Handle<v8::Object> Exports);
@@ -926,6 +947,7 @@ class TNodeJsStore : public node::ObjectWrap {
 private:
     // Node framework
     static v8::Persistent<v8::Function> Constructor;
+    ~TNodeJsStore() { TNodeJsUtil::ObjNameH.GetDat(GetClassId()).Val3++; TNodeJsUtil::ObjCount.Val3++; }
 public: 
     // Node framework 
     static void Init(v8::Handle<v8::Object> exports);
@@ -1662,6 +1684,7 @@ class TNodeJsRec: public node::ObjectWrap {
 private:
     // Modified node framework: one record template per each base,storeId combination
     static TVec<TVec<v8::Persistent<v8::Function> > > BaseStoreIdConstructor;
+    ~TNodeJsRec() { TNodeJsUtil::ObjNameH.GetDat(GetClassId()).Val3++; TNodeJsUtil::ObjCount.Val3++; }
 public:
     // Node framework 
     static void Init(const TWPt<TQm::TStore>& Store);
@@ -1879,6 +1902,7 @@ class TNodeJsRecByValV: public node::ObjectWrap {
 private:
     // Node framework
     static v8::Persistent<v8::Function> Constructor;
+    ~TNodeJsRecByValV() { TNodeJsUtil::ObjNameH.GetDat(GetClassId()).Val3++; TNodeJsUtil::ObjCount.Val3++; }
 public:
     // Node framework
     static void Init(v8::Handle<v8::Object> Exports);
@@ -1985,6 +2009,7 @@ class TNodeJsRecSet: public node::ObjectWrap {
 private:
     // Node framework
     static v8::Persistent<v8::Function> Constructor;
+    ~TNodeJsRecSet() { TNodeJsUtil::ObjNameH.GetDat(GetClassId()).Val3++; TNodeJsUtil::ObjCount.Val3++; }
 public:
     // Node framework 
     static void Init(v8::Handle<v8::Object> exports);
@@ -2879,6 +2904,7 @@ class TNodeJsStoreIter: public node::ObjectWrap {
 private:
     // Node framework
     static v8::Persistent<v8::Function> Constructor;
+
 public:
     // Node framework 
     static void Init(v8::Handle<v8::Object> exports);
@@ -2900,7 +2926,7 @@ public:
 public:
     
     // delete placeholder
-    ~TNodeJsStoreIter() { RecObj.Reset(); }
+    ~TNodeJsStoreIter() { TNodeJsUtil::ObjNameH.GetDat(GetClassId()).Val3++; TNodeJsUtil::ObjCount.Val3++; RecObj.Reset(); }
     
     /**
     * Moves to the next record.
@@ -2992,6 +3018,7 @@ class TNodeJsIndexKey: public node::ObjectWrap {
 private:
     // Node framework
     static v8::Persistent<v8::Function> Constructor;
+    ~TNodeJsIndexKey() { TNodeJsUtil::ObjNameH.GetDat(GetClassId()).Val3++; TNodeJsUtil::ObjCount.Val3++; }
 public:
     // Node framework
     static void Init(v8::Handle<v8::Object> exports);
@@ -3133,7 +3160,7 @@ public:
 * @typedef {Object} FeatureExtractorNumeric
 * The feature extractor of type `'numeric'`. Used for constructing {@link module:qm.FeatureSpace} objects.
 * @property {string} type - The type of the extractor. <b>Important</b>: It must be equal `'numeric'`.
-* @property {boolean} [normalize = 'false'] - Normalize values between 0.0 and 1.0.
+* @property {string} [normalize = 'none'] - Normalize values between 0.0 and 1.0 if set to `'scale'`. Standardize values to mean = 0 and sdiv = 1 if set to `'var'`.
 * @property {number} [min] - The minimal value used to form the normalization.
 * @property {number} [max] - The maximal value used to form the normalization.
 * @property {string} field - The name of the field from which to take the value.
@@ -3153,7 +3180,7 @@ public:
 * });
 * // create a feature space containing the numeric extractor, where the values are
 * // normalized, the values are taken from the field "Grade"
-* var ftr = new qm.FeatureSpace(base, { type: "numeric", source: "Class", normalize: true, field: "Grade" });
+* var ftr = new qm.FeatureSpace(base, { type: "numeric", source: "Class", normalize: "scale", field: "Grade" });
 * base.close();
 */
 
@@ -3428,6 +3455,7 @@ class TNodeJsFtrSpace : public node::ObjectWrap {
     friend class TNodeJsUtil;
 private:
     static v8::Persistent<v8::Function> Constructor;
+    ~TNodeJsFtrSpace() { TNodeJsUtil::ObjNameH.GetDat(GetClassId()).Val3++; TNodeJsUtil::ObjCount.Val3++; }
 public:
     // Node framework
     static void Init(v8::Handle<v8::Object> exports);